requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/dev.txt --pre requirements.in/dev.txt
#
-e git+https://github.com/thisch/pytest-sphinx.git#egg=pytest-sphinx
    # via -r requirements.in/test.txt
absl-py==0.12.0
    # via
    #   tensorflow-datasets
    #   tensorflow-metadata
alabaster==0.7.12
    # via sphinx
appdirs==1.4.4
    # via
    #   black
    #   tox
    #   virtualenv
astroid==2.5.1
    # via pylint
attrs==20.3.0
    # via
    #   flake8-bugbear
    #   pytest
    #   tensorflow-datasets
babel==2.9.0
    # via sphinx
black==20.8b1
    # via -r requirements.in/linting.txt
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via
    #   doc8
    #   requests
click==7.1.2
    # via
    #   -r requirements.in/base.txt
    #   black
    #   sphinx-click
codecov==2.1.11
    # via -r requirements.in/test.txt
colorama==0.4.4
    # via
    #   sphinx-autobuild
    #   tox
coverage==5.5
    # via
    #   -r requirements.in/test.txt
    #   codecov
    #   pytest-cov
dill==0.3.3
    # via tensorflow-datasets
distlib==0.3.1
    # via virtualenv
doc8==0.8.1
    # via -r requirements.in/docs.txt
docutils==0.17b1
    # via
    #   doc8
    #   restructuredtext-lint
    #   sphinx
    #   sphinx-click
filelock==3.0.12
    # via
    #   tox
    #   virtualenv
flake8-bugbear==21.3.2
    # via -r requirements.in/linting.txt
flake8==3.9.0
    # via
    #   -r requirements.in/linting.txt
    #   flake8-bugbear
future==0.18.2
    # via tensorflow-datasets
googleapis-common-protos==1.53.0
    # via tensorflow-metadata
idna==2.10
    # via requests
imagesize==1.2.0
    # via sphinx
iniconfig==1.1.1
    # via pytest
isort[pyproject,requirements]==5.8.0
    # via
    #   -r requirements.in/linting.txt
    #   pylint
jinja2==3.0.0a1
    # via sphinx
lazy-object-proxy==1.6.0
    # via astroid
livereload==2.6.3
    # via sphinx-autobuild
markupsafe==2.0.0rc1
    # via jinja2
mccabe==0.6.1
    # via
    #   flake8
    #   pylint
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
mypy==0.812
    # via -r requirements.in/linting.txt
numpy==1.20.1
    # via tensorflow-datasets
packaging==20.9
    # via
    #   pytest
    #   sphinx
    #   tox
pathspec==0.8.1
    # via black
pbr==5.5.1
    # via stevedore
pluggy==1.0.0.dev0
    # via
    #   pytest
    #   tox
promise==2.3
    # via tensorflow-datasets
protobuf==3.15.6
    # via
    #   googleapis-common-protos
    #   tensorflow-datasets
    #   tensorflow-metadata
py==1.10.0
    # via pytest
pycodestyle==2.7.0
    # via flake8
pydocstyle==6.0.0
    # via -r requirements.in/docs.txt
pyflakes==2.3.0
    # via flake8
pygments==2.8.1
    # via
    #   doc8
    #   sphinx
pylint==2.7.2
    # via -r requirements.in/linting.txt
pyparsing==3.0.0b2
    # via packaging
pytest-cov==2.11.1
    # via -r requirements.in/test.txt
pytest==6.2.2
    # via
    #   -r requirements.in/test.txt
    #   pytest-cov
    #   pytest-sphinx
pytz==2021.1
    # via babel
regex==2021.3.17
    # via black
requests==2.25.1
    # via
    #   codecov
    #   sphinx
    #   tensorflow-datasets
restructuredtext-lint==1.3.2
    # via doc8
rope==0.18.0
    # via -r requirements.in/dev.txt
six==1.15.0
    # via
    #   absl-py
    #   doc8
    #   livereload
    #   promise
    #   protobuf
    #   tensorflow-datasets
    #   virtualenv
snowballstemmer==2.1.0
    # via
    #   pydocstyle
    #   sphinx
sphinx-autobuild==2021.3.14
    # via -r requirements.in/docs.txt
sphinx-autodoc-typehints==1.11.1
    # via -r requirements.in/docs.txt
sphinx-click==2.7.1
    # via -r requirements.in/docs.txt
sphinx-copybutton==0.3.1
    # via -r requirements.in/docs.txt
sphinx-rtd-theme==0.5.1
    # via -r requirements.in/docs.txt
sphinx==3.5.3
    # via
    #   -r requirements.in/docs.txt
    #   sphinx-autobuild
    #   sphinx-autodoc-typehints
    #   sphinx-click
    #   sphinx-copybutton
    #   sphinx-rtd-theme
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==1.0.3
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.4
    # via sphinx
stevedore==3.3.0
    # via doc8
tensorflow-datasets==4.2.0
    # via -r requirements.in/base.txt
tensorflow-metadata==0.29.0
    # via tensorflow-datasets
termcolor==1.1.0
    # via tensorflow-datasets
toml==0.10.2
    # via
    #   black
    #   pylint
    #   pytest
    #   tox
tornado==6.1
    # via livereload
tox==4.0.0a6
    # via -r requirements.in/test.txt
tqdm==4.59.0
    # via tensorflow-datasets
typed-ast==1.4.2
    # via
    #   black
    #   mypy
typing-extensions==3.7.4.3
    # via
    #   black
    #   mypy
urllib3==1.26.4
    # via requests
virtualenv==20.4.3
    # via tox
wrapt==1.12.1
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
